+++ allegrocheckin/allegrocheckin/lib/catalog_products/catalog_products.dart
@@ -2,6 +2,7 @@ import 'package:allegrocheckin/Service/ReserveService.dart';
 import 'package:allegrocheckin/models/categorias.dart';
 import 'package:allegrocheckin/models/commandresult_model.dart';
 import 'package:allegrocheckin/models/products.dart';
+import 'package:confirm_dialog/confirm_dialog.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -83,11 +84,14 @@ class _ProductListPageState extends State<ProductListPage> {
                   trailing: IconButton(
                     icon: Icon(Icons.delete),
                     onPressed: () async {
-                      await ReserveService()
-                          .deleteProduct(productList[index].id);
-                      setState(() {
-                        var i = 1;
-                      });
+                      if (await confirm(context)) {
+                        await ReserveService()
+                            .deleteProduct(productList[index].id);
+                        setState(() {
+                          var i = 1;
+                        });
+                      }
+                      return print('pressedCancel');
                     },
                   ),
                 );
